catagory updated

+++ src/main/java/com/crudoperation/jw/repo/CatagorieRepository.java
@@ -3,5 +3,5 @@ package com.crudoperation.jw.repo;
 import com.crudoperation.jw.model.Catagorie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CatagorieRepository extends JpaRepository<Catagorie, Long> {
+public interface CatagorieRepository extends JpaRepository<Catagorie,Integer> {
 }
